add data for 'Force on Lid'

+++ 1/code/run.sh
@@ -11,40 +11,46 @@ echo "-\| INITIALIZE"
 cd param/
 for run in $(echo */)
 do
-    echo "-\ Copying Files..."
-    cp -r "$run"* ../sim
-    echo "-\ CD"
-    cd ../sim/
-    echo "-\ Staging"
-    python ../analysis/sample.py
-    echo "-\ Allclean"
-    ./Allclean
-    echo "-\ Allrun"
-    { time ./Allrun 2>> time.txt ; } 
-    echo "-\ Simulation Complete."
-    echo "-\ Post-processing via Sample..."
-    postProcess -func sample
-    echo "-\ Complete."
-    echo "-\ Copying Data..."
-    for step in $(ls -d *[0-9]*)
-    do
-        mkdir -p ../data/"$run"
-        mv -f "$step"/ ../data/"$run"
-    done
-    mv -f constant/ ../data/"$run"
-    mv -f postProcessing/ ../data/"$run"
-    echo "-\ Sweeping Dirt ..."
-    # ./Allclean
-    rm -r system/
-    rm README.md
-    rm sets.txt
-    echo "-\ CD"
-    cd ../data/"$run"/
-    echo "-\ Creating _.Foam ..."
-    touch _.foam
-    echo "-\ CD"
-    cd ../../param/
-    echo "-/| DONE |"
+    echo "Data Exists?..."
+    if [[ -d "../data/$run" ]]
+    then
+        echo "Data Exists. Skipping Run $run."
+    else
+        echo "-\ Copying Files..."
+        cp -r "$run"* ../sim
+        echo "-\ CD"
+        cd ../sim/
+        echo "-\ Staging"
+        python ../analysis/sample.py
+        echo "-\ Allclean"
+        ./Allclean
+        echo "-\ Allrun"
+        { time ./Allrun 2>> time.txt ; } 
+        echo "-\ Simulation Complete."
+        echo "-\ Post-processing via Sample..."
+        postProcess -func sample
+        echo "-\ Complete."
+        echo "-\ Copying Data..."
+        for step in $(ls -d *[0-9]*)
+        do
+            mkdir -p ../data/"$run"
+            mv -f "$step"/ ../data/"$run"
+        done
+        mv constant/ ../data/"$run"
+        mv postProcessing/ ../data/"$run"
+        echo "-\ Sweeping Dirt ..."
+        # ./Allclean
+        rm -r system/
+        rm README.md
+        rm sets.txt
+        echo "-\ CD"
+        cd ../data/"$run"/
+        echo "-\ Creating _.Foam ..."
+        touch _.foam
+        echo "-\ CD"
+        cd ../../param/
+        echo "-/| DONE |"
+    fi
 done
 echo "-\| TIMING Data |"
 mkdir -p ../data/time/
